feat: 실행 모듈, 테스트 추가
