--- app/src/main/java/java/calculator/App.java
@@ -1,14 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.calculator;
-
-public class App {
-    public String getGreeting() {
-        return "Hello World!";
-    }
-
-    public static void main(String[] args) {
-        System.out.println(new App().getGreeting());
-    }
-}
