--- app/src/main/java/java/calculator/common/Calculator.java
@@ -1,10 +1,0 @@
-package java.calculator.common;
-
-import java.calculator.common.engine.CalculatorEngine;
-
-public class Calculator {
-
-    public int calculate(String rawExpression) {
-        return new CalculatorEngine(rawExpression).evaluate();
-    }
-}
