--- app/src/main/java/java/calculator/common/command/Command.java
@@ -1,5 +1,0 @@
-package java.calculator.common.command;
-
-public interface Command {
-    void execute();
-}
