--- app/src/main/java/java/calculator/common/reader/Reader.java
@@ -1,5 +1,0 @@
-package java.calculator.common.reader;
-
-public interface Reader {
-    String readLine();
-}
