--- app/src/main/java/java/calculator/common/utils/Tokenizer.java
@@ -1,15 +1,0 @@
-package java.calculator.common.utils;
-
-public class Tokenizer {
-    private final String[] tokens;
-
-    public Tokenizer(String expression) {
-        this.tokens = expression.split(" ");
-    }
-
-    public String[] getTokens() {
-        return this.tokens;
-    }
-}
-
-
