--- app/src/main/java/java/calculator/common/writer/Writer.java
@@ -1,5 +1,0 @@
-package java.calculator.common.writer;
-
-public interface Writer {
-    void write(String message);
-}
