--- app/src/main/java/java/calculator/console/reader/ConsoleReader.java
@@ -1,18 +1,0 @@
-package java.calculator.console.reader;
-
-import java.calculator.common.reader.Reader;
-import java.util.Scanner;
-
-public class ConsoleReader implements Reader {
-    private final Scanner scanner;
-
-    public ConsoleReader() {
-        this.scanner = new Scanner(System.in);
-    }
-
-    @Override
-    public String readLine() {
-        return scanner.nextLine();
-    }
-}
-
