--- app/src/main/java/java/calculator/console/writer/ConsoleWriter.java
@@ -1,11 +1,0 @@
-package java.calculator.console.writer;
-
-import java.calculator.common.writer.Writer;
-
-public class ConsoleWriter implements Writer {
-
-    @Override
-    public void write(String message) {
-        System.out.println(message);
-    }
-}
