+++ app/src/main/java/programmers/java/calulator/common/command/Command.java
@@ -1,0 +1,5 @@
+package programmers.java.calulator.common.command;
+
+public interface Command {
+    void execute();
+}
