+++ app/src/main/java/programmers/java/calulator/common/reader/Reader.java
@@ -1,0 +1,5 @@
+package programmers.java.calulator.common.reader;
+
+public interface Reader {
+    String readLine();
+}
