+++ app/src/main/java/programmers/java/calulator/common/runner/ApplicationRunner.java
@@ -1,0 +1,5 @@
+package programmers.java.calulator.common.runner;
+
+public interface ApplicationRunner {
+    void run();
+}
