+++ app/src/main/java/programmers/java/calulator/common/utils/Tokenizer.java
@@ -1,0 +1,15 @@
+package programmers.java.calulator.common.utils;
+
+public class Tokenizer {
+    private final String[] tokens;
+
+    public Tokenizer(String expression) {
+        this.tokens = expression.split(" ");
+    }
+
+    public String[] getTokens() {
+        return this.tokens;
+    }
+}
+
+
