+++ app/src/main/java/programmers/java/calulator/common/writer/Writer.java
@@ -1,0 +1,5 @@
+package programmers.java.calulator.common.writer;
+
+public interface Writer {
+    void write(String message);
+}
