+++ app/src/main/java/programmers/java/calulator/console/reader/ConsoleReader.java
@@ -1,0 +1,19 @@
+package programmers.java.calulator.console.reader;
+
+import programmers.java.calulator.common.reader.Reader;
+
+import java.util.Scanner;
+
+public class ConsoleReader implements Reader {
+    private final Scanner scanner;
+
+    public ConsoleReader() {
+        this.scanner = new Scanner(System.in);
+    }
+
+    @Override
+    public String readLine() {
+        return scanner.nextLine();
+    }
+}
+
