--- app/src/test/java/java/calculator/AppTest.java
@@ -1,14 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package java.calculator;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class AppTest {
-    @Test public void appHasAGreeting() {
-        App classUnderTest = new App();
-        assertNotNull("app should have a greeting", classUnderTest.getGreeting());
-    }
-}
