+++ app/src/test/java/programmers/java/calulator/AppTest.java
@@ -1,0 +1,9 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package programmers.java.calulator;
+
+
+public class AppTest {
+
+}
